Changes to include more categories + reduce time consuption of analyzers (#96)
- Analysis/include/AnaTuple.h: Created a boolean variable runKinFit, in the constructor of AnaTupleWriter and added the variables of gen_match_1 and gen_match_1 to the variable list

- Analysis/include/AnalysisCategories.h: Implementation of a new function "Contains", called BaseEventAnalyzer, to discriminate if an event contains or not the category pass in the sources

- Analysis/include/BaseEventAnalyzer.h: Changes in the DetermineEventCategories to use new function Contains, and fix some miscalculated values. 

 - Analysis/include/EventAnalyzerCore.h:  Added boolean argument runKinFit

 - Instruments/scripts/dataIdCalculator.py: New script that allows to get the dataId number for a given syncDataIds

+++ Studies/source/DY_estimation.cxx
@@ -262,7 +262,7 @@ public:
                             Components((static_cast<std::string>("expdf_")+ToString(cat)+"_"+ToString(sub_cat)+
                                         static_cast<std::string>("_other_bkg_muMu")).c_str() ),
                             ProjWData(rooCategories,combData),LineStyle(kDashed)) ;
-                gPad->SetLeftMargin(0.15) ; frame->GetYaxis()->SetTitleOffset(static_cast<float>(1.4)) ; frame->Draw() ;
+                gPad->SetLeftMargin(0.15f) ; frame->GetYaxis()->SetTitleOffset(1.4f) ; frame->Draw() ;
                 c->Write();
             }
         }
@@ -286,4 +286,3 @@ private:
 
 } // namesapce analysis
 PROGRAM_MAIN(analysis::DY_estimation, Arguments) // definition of the main program function
-
